requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==24.10.0
click==8.1.7
decorator==5.1.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
flake8==7.1.1
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
